Fix the wrong rotation of the image views, #70 . Fixes labeling on x,y.
To do this, introduce 'data coordinates'. These are the t,x,y,z,c
coordinates. In imageView2D, we use these data coordinates and know
how to convert back and forth from scene coordinates using a QTransform.

This also fixes the labeling on the x and y slices.

To close #70, the coordinate systems need to be formally defined,
documented and the code in slicingpositionmarker adapted. Furhtermore,
it would be good if we can get different scene transformations to work.