prevent rendering artefacts of tiles at certain zoom levels
By letting neighbouring patches overlap slightly, rendering artefacts
can be prevented. However, it is necessary to differentiate between
the bounding box of the enlarged image and the bounding box of the
underlying patch. The latter bounding boxes should not overlap, and
should be used when querying which patches are visible in the current
viewport. Otherwise, too many patches could be requested.